requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.11.17
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.8
daphne==4.1.2
distro==1.8.0
dj-database-url==2.2.0
dj-rest-auth==6.0.0
Django==4.2.13
django-allauth==0.63.3
django-cors-headers==4.4.0
django-filter==24.3
django-otp==1.5.1
django-otp-twilio==1.0.5
django-phonenumber-field==8.0.0
django-redis==5.4.0
django-twilio==0.14.3.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
exceptiongroup==1.2.1
Faker==21.0.0
frozenlist==1.4.1
h11==0.14.0
hiredis==3.0.0
httpcore==1.0.2
httpx==0.25.2
hyperlink==21.0.0
idna==3.6
incremental==24.7.2
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
multidict==6.0.5
oauthlib==3.2.2
openai==1.5.0
packaging==24.1
phonenumbers==8.13.40
pillow==10.4.0
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
pyOpenSSL==24.2.1
pyotp==2.9.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.8
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.18.1
service-identity==24.1.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tomli==2.0.1
tqdm==4.66.1
twilio==9.2.3
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.2.2
yarl==1.9.4
zope.interface==6.4.post2